move dial queue to dedicated file; some fixes.
